requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.9
pymysql==1.0.3
python-dotenv==1.0.0
bcrypt==4.0.1
python-jose==3.3.0
python-multipart==0.0.6
pyjwt==2.6.0
cryptography==40.0.2
aiofiles==23.1.0
psycopg2-binary==2.9.6